fix mtr bus database

+++ app/src/main/java/com/alvinhkh/buseta/mtr/MtrBusWorker.kt
@@ -4,6 +4,7 @@ import android.content.Context
 import android.util.SparseArray
 import android.util.SparseIntArray
 import androidx.core.util.set
+import androidx.preference.PreferenceManager
 import androidx.room.Room
 import androidx.work.Data
 import androidx.work.Worker
@@ -20,6 +21,7 @@ import com.alvinhkh.buseta.search.dao.SuggestionDatabase
 import com.alvinhkh.buseta.search.model.Suggestion
 import com.google.firebase.remoteconfig.FirebaseRemoteConfig
 import timber.log.Timber
+import java.io.File
 import java.lang.Exception
 
 class MtrBusWorker(context : Context, params : WorkerParameters)
@@ -30,6 +32,8 @@ class MtrBusWorker(context : Context, params : WorkerParameters)
     private val suggestionDatabase = SuggestionDatabase.getInstance(context)
 
     private val firebaseRemoteConfig = FirebaseRemoteConfig.getInstance()
+
+    private val preferences = PreferenceManager.getDefaultSharedPreferences(context)
 
     override fun doWork(): Result {
         val manualUpdate = inputData.getBoolean(C.EXTRA.MANUAL, false)
@@ -44,14 +48,19 @@ class MtrBusWorker(context : Context, params : WorkerParameters)
         val timeNow = System.currentTimeMillis() / 1000
 
         val busDatabaseFileName = firebaseRemoteConfig.getString("mtr_bus_database_file")
+        val lastBusDatabaseFileName = preferences.getString("mtr_bus_database_file", "")?: ""
         if (busDatabaseFileName.isEmpty()) {
             return Result.failure(outputData)
+        }
+        if (!manualUpdate && lastBusDatabaseFileName == busDatabaseFileName) {
+            Timber.d("MTR Bus database: %s", busDatabaseFileName)
+            return Result.success(outputData)
         }
 
         applicationContext.deleteDatabase("E_Bus.db")
         val database = Room.databaseBuilder(applicationContext, MtrBusDatabase::class.java, "E_Bus.db")
                 .allowMainThreadQueries()
-                .createFromAsset("database/$busDatabaseFileName")
+                .createFromFile(File("${applicationContext.cacheDir}/$busDatabaseFileName"))
                 .build()
 
         val mtrBusFareArray = SparseArray<MtrBusFare>()
@@ -163,6 +172,10 @@ class MtrBusWorker(context : Context, params : WorkerParameters)
 
         routeDatabase?.routeDao()?.delete(C.PROVIDER.AESBUS, timeNow)
         routeDatabase?.routeStopDao()?.delete(C.PROVIDER.AESBUS, timeNow)
+
+        val editor = preferences.edit()
+        editor.putString("mtr_bus_database_file", busDatabaseFileName)
+        editor.apply()
 
         return Result.success(outputData)
     }
